Chrome complains that the PDF is broken
When the HTML includes heading tags, wkhtmltopdf would try and parse
these tags to setup a table of contents for the PDF. However this
caused chrome to complain that the PDF was broken when rendering
the PDF using the built in PDF viewer. by adding the --no-outline
option to the command wkhtmltopdf wont try and do this, thus chrome
wont complain.

The side affect of this is that we know require the latest version of
wkhtmltopdf since it's the only one with this feature